Extract types to convex lib
